Separated king movement to a different file
